pyproject.toml
[project]
name = "onelife"
version = "0.1.0"
description = "A personal newsfeed aggregator."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "feedparser>=6.0.12",
    "requests>=2.32.5",
    "beautifulsoup4>=4.14.2",
    "pyyaml>=6.0.3",
    "python-dateutil>=2.9.0.post0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
